Fix Gradle 3.3 compatibility by moving build service to own class
